adjust file structure and naming
