+++ backend/api/Features/Cases/CaseComparison/CaseComparisonService.cs
@@ -1,8 +1,8 @@
-using api.Features.Cases.Recalculation.Types.TotalProductionVolume;
 using api.Features.Profiles;
 using api.Features.Profiles.Dtos;
 using api.Features.Profiles.TimeSeriesMerging;
 using api.Models;
+
 using static api.Features.Profiles.VolumeConstants;
 
 namespace api.Features.Cases.CaseComparison;
