+++ backend/tests/Features/Cases/Recalculation/Calculators/CalculateNpvServiceTests.cs
@@ -1,4 +1,4 @@
-using api.Features.Cases.Recalculation.Calculators.CalculateNpv;
+using api.Features.Cases.Recalculation.RevenuesAndCashflow;
 using api.Features.Profiles;
 using api.Models;
