+++ backend/tests/Features/Cases/Recalculation/Calculators/EconomicsHelperTests.cs
@@ -1,4 +1,4 @@
-using api.Features.Cases.Recalculation.Calculators.Helpers;
+using api.Features.Cases.Recalculation.Helpers;
 using api.Features.Profiles.Dtos;
 
 using Xunit;
