+++ backend/tests/Features/Cases/Recalculation/Calculators/EmissionCalculationHelperTests.cs
@@ -1,4 +1,4 @@
-using api.Features.Cases.Recalculation.Types.Helpers;
+using api.Features.Cases.Recalculation.Helpers;
 using api.Features.Profiles;
 using api.Features.Profiles.Dtos;
 using api.Features.Profiles.TimeSeriesMerging;
