+++ backend/tests/Features/Cases/Recalculation/Calculators/OpexCostProfile/CalculateWellInterventionCostProfileTests.cs
@@ -1,4 +1,4 @@
-using api.Features.Cases.Recalculation.Types.OpexCostProfile;
+using api.Features.Cases.Recalculation.Cost;
 using api.Features.Profiles;
 using api.Models;
 using api.Models.Enums;
