+++ backend/tests/Features/Cases/Recalculation/Calculators/StudyCostProfileServiceTests.cs
@@ -1,4 +1,4 @@
-using api.Features.Cases.Recalculation.Types.StudyCostProfile;
+using api.Features.Cases.Recalculation.Cost;
 using api.Features.Profiles;
 using api.Models;
