+++ backend/tests/Features/Cases/Recalculation/Types/RigCostProfileServiceTests.cs
@@ -1,4 +1,4 @@
-using api.Features.Cases.Recalculation.Types.RigCostProfile;
+using api.Features.Cases.Recalculation.Cost;
 using api.Features.Profiles;
 using api.Models;
 using api.Models.Enums;
